Arrange FlatArray From factory test files
